backend/requirements.txt
# NeuralStark Backend Dependencies - Python 3.13+ Compatible
# Core FastAPI and Server
fastapi>=0.110.0
uvicorn[standard]>=0.25.0
python-multipart>=0.0.9
python-dotenv>=1.0.0

# Database
motor>=3.3.0
pymongo>=4.5.0
dnspython>=2.6.0

# AI/ML - LLM Integration
cerebras_cloud_sdk>=1.0.0

# Vector Store and Embeddings
chromadb>=0.4.0
overrides>=7.0.0
sentence-transformers>=2.5.0

# RAG Framework
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1

# Document Processing
pypdf>=4.0.0
pdfplumber>=0.10.0
pdfminer.six>=20221105
pdf2image>=1.17.0
python-docx>=1.1.0
openpyxl>=3.1.0
odfpy>=1.4.1
lxml>=4.9.0

# OCR
pytesseract>=0.3.10
pillow>=10.0.0

# File System Monitoring
watchdog>=4.0.0

# HTTP and Async
httpx>=0.27.0
aiohttp>=3.9.0
anyio>=4.0.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
tenacity>=8.2.0
tqdm>=4.66.0

# Data Processing
numpy>=1.26.0
pandas>=2.1.0

# Testing (optional but recommended)
pytest>=8.0.0
httpx>=0.27.0
defusedxml>=0.7.0
pybase64>=1.4.0
jsonschema>=4.20.0
pyyaml>=6.0.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.3.0
distro>=1.8.0
posthog>=6.0.0

# RAG Enhancement Libraries
rank-bm25>=0.2.2
pyspellchecker>=0.8.1
nltk>=3.8.1
textdistance>=4.6.0
rapidfuzz>=3.6.0

# French NLP Support
spacy>=3.7.0

# French NLP Support
spacy>=3.7.0
